requirements.txt
altgraph==0.17.4
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
importlib-metadata==8.7.0
natsort==8.4.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pymupdf==1.26.0
pypdf2==3.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
soupsieve==2.7
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.4.0
zipp==3.21.0
